Move files into CgmesConformity1ModifiedCatalog.
