+++ cgmes/cgmes-conformity/src/main/resources/conformity-modified/cas-1.1.3-data-4.0.3/SmallGrid/BusBranch_busbarSections_ipMax/SmallGridTestConfiguration_BC_EQ_v3.0.0.xml
